+++ src/database/migrations/2024_01_20_115711_create_group_users_table.php
@@ -13,6 +13,13 @@ return new class extends Migration
     {
         Schema::create('group_users', function (Blueprint $table) {
             $table->id();
+            $table->string('status', 25);   // 'approved', 'pending'
+            $table->string('role', 25);
+            $table->string('token', 1024)->nullable();
+            $table->timestamp('token_expire_date')->nullable();
+            $table->timestamp('token_used')->nullable();
+            $table->foreignId('user_id')->constrained('users');
+            $table->foreignId('group_id')->constrained('groups');
             $table->timestamps();
         });
     }