+++ src/database/migrations/2024_01_20_115740_create_posts_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
             $table->longText('body')->nullable();
-            $table->foreignId('created_by')->constrained('user');
-            $table->foreignId('updated_by')->constrained('user');
-            $table->foreignId('deleted_by')->nullable()->constrained('user');
+            $table->foreignId('user_id')->constrained('users');
+            $table->foreignId('group_id')->constrained('groups');
+            $table->foreignId('deleted_by')->nullable()->constrained('users');
             $table->timestamp('deleted_at')->nullable();
             $table->timestamps();
         });